backend/requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain-core
langchain
chromadb
mem0ai
networkx
python-dotenv
requests
litellm # good for model abstraction
